Task 10
Added getwalletversion
